Reflected the discussion in http://togetter.com/li/267335.
Added linking tests on duplicate symbols defined in text sections. Any
combination of kinds of object/library files is tested, i.e., for
combinations of .o and .o, .o and .a, .o and .so, .a and .o, .a and .a, .a
and .so, .so and .o, .so and .a, .so and .so. For each of these combinations,
linking tests both with and without `--nodup' have been tested. For this set
of tests, `bjam toolset=gcc-4.4.6,gcc-4.5.3,gcc-4.6.2,gcc-4.7.0 variant=debug,release link=static,shared threading=single,multi' has been passed, i.e.,
9*2*4*2*2*2=576 tests have been passed totally.